requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
faker
gunicorn
numpy
geopy
pandas
scikit-learn
python-multipart
email-validator
pydantic-settings>=2.0.0
pydantic-extra-types>=2.0.0
markdown>=3.4.1
pygments>=2.13.0